Move testQuery to core-ui, add test coverage for Queries in columnChoices
