+++ superset-frontend/packages/superset-ui-chart-controls/test/utils/columnChoices.test.tsx
@@ -16,11 +16,11 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { DatasourceType } from '@superset-ui/core';
+import { DatasourceType, QueryResponse, testQuery } from '@superset-ui/core';
 import { columnChoices } from '../../src';
 
 describe('columnChoices()', () => {
-  it('should convert columns to choices', () => {
+  it('should convert columns to choices when source is a Dataset', () => {
     expect(
       columnChoices({
         id: 1,
@@ -55,5 +55,14 @@ describe('columnChoices()', () => {
 
   it('should return empty array when no columns', () => {
     expect(columnChoices(undefined)).toEqual([]);
+  });
+
+  it('should convert columns to choices when source is a Query', () => {
+    expect(columnChoices(testQuery as QueryResponse)).toEqual([
+      ['Column 1', 'Column 1'],
+      ['Column 2', 'Column 2'],
+      ['Column 3', 'Column 3'],
+    ]);
+    expect.anything();
   });
 });
