+++ superset-frontend/src/SqlLab/fixtures.ts
@@ -19,7 +19,6 @@
 import sinon from 'sinon';
 import * as actions from 'src/SqlLab/actions/sqlLab';
 import { ColumnKeyTypeType } from 'src/SqlLab/components/ColumnElement';
-import { DatasourceType, Query } from '@superset-ui/core';
 
 export const mockedActions = sinon.stub({ ...actions });
 
@@ -572,42 +571,4 @@ export const query = {
   runAsync: false,
   ctas: false,
   cached: false,
-};
-
-export const testQuery: Query = {
-  id: 'clientId2353',
-  dbId: 1,
-  sql: 'SELECT * FROM something',
-  sqlEditorId: defaultQueryEditor.id,
-  tab: 'unimportant',
-  tempTable: '',
-  ctas: false,
-  cached: false,
-  errorMessage: null,
-  extra: { progress: null },
-  isDataPreview: false,
-  progress: 0,
-  resultsKey: null,
-  state: 'success',
-  tempSchema: null,
-  trackingUrl: null,
-  templateParams: null,
-  rows: 42,
-  queryLimit: 100,
-  limitingFactor: '',
-  endDttm: 1476910579693,
-  duration: '',
-  startDttm: 1476910566092.96,
-  time: {},
-  user: {},
-  userId: 1,
-  db: {},
-  started: '',
-  querylink: {},
-  queryId: 1,
-  executedSql: '',
-  output: '',
-  actions: {},
-  type: DatasourceType.Query,
-  columns: [],
 };
